[feat] Lotto - 자동번호 구현 및 매칭 구현
